Fix "Content-disposition" KeyError
If we go to download a file of a weird type (e.g. a very short movie) and cannot find a "Content-Disposition:" header, we may throw a "KeyError: Content-Disposition".

What we wanted to do was raise a "cannot guess extension" exception and print the URL. We can do that by checking whether the header exists rather than checking whether the value of the header is true.